separate tests no assert
