+++ .locust.conf
@@ -1,5 +1,5 @@
 # master.conf in current directory
-locustfile = tests/locust/test_under_rate_limit.py
+locustfile = tests/locust/test_over_quota.py
 headless = true
 host = https://internal-dev-sandbox.api.service.nhs.uk/comms-pr-171
 html = report.html
