+++ Makefile
@@ -79,7 +79,12 @@ PROD_TEST_CMD := $(TEST_CMD) \
 	(rm -rf node_modules; npm install --legacy-peer-deps; npm run sandbox-postman-collection)
 
 .run-locust-tests:
-	(poetry run locust -f tests/locust/locustfile.py --headless -u 100 -r 5 -t 1m --html locust-report.html)
+	(sleep 60; \
+	poetry run locust -f tests/locust/test_no_errors.py --config .locust.conf; \
+	sleep 60; \
+	poetry run locust -f tests/locust/test_over_quota.py --config .locust.conf; \
+	sleep 60; \
+	poetry run locust -f tests/locust/test_over_spike_arrest.py --config .locust.conf;)
 
 #Command to run end-to-end smoketests post-deployment to verify the environment is working
 smoketest:
